cli/command/service: remove AppendServiceStatus (API <v1.41)
This function was added in 7405ac5c2d63c5db909a6c5093bb4304f376d9af as
a fallback for API < v1.41, which did not include the service status
in the response. Current API versions return this information, so there's
no need to fetch it manually.

It was not gated by API version for some tests (which didn't set API
version), but should not be needed for non-test situations.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>